Fetaure: NavHostController supports NavigationResults & Feature: Schedules - Add ScheduleDetail to add/edit them

+++ modules/library-navigation-api/src/commonMain/kotlin/com/jarroyo/library/navigation/api/destination/NavigationDestination.kt
@@ -15,7 +15,7 @@ abstract class NavigationDestination {
 
         companion object {
             val uniqueId
-                get() = Random.nextInt()
+                get() = Random.nextLong()
         }
     }
